+++ 0x03-more_functions_nested_loops/holberton.h
@@ -1,6 +1,6 @@
 int mul(int a, int b);
 int _putchar(char c);
-long _pow(int, int);
+int _pow(int, int);
 int _abs(int n);
 int _isupper(int c);
 int _isdigit(int c);
